requirements.txt
textual>=0.50.0
anthropic>=0.30.0
networkx>=3.1
python-frontmatter>=1.0.0
PyYAML>=6.0
Pillow>=10.0.0
pyte>=0.8.0
ptyprocess>=0.7.0
